save embeddings working
