+++ python_search/llm_next_item_predictor/t5/t5_embeddings.py
@@ -1,22 +1,38 @@
-#!/usr/bin/env python3
-
 from typing import List
 
 import pandas as pd
-from sklearn.metrics.pairwise import cosine_similarity
 
 from python_search.configuration.loader import ConfigurationLoader
+from python_search.llm_next_item_predictor.t5.config import T5Model
 from python_search.semantic_search.distilbert import to_embedding2, to_embedding
 
-class SemanticSearch:
+class T5Embeddings:
     """
     Build entry embeddings and save them
     """
 
-    pickled_location = "~/.python_search/bert_entry_embeddings.pkl"
+    pickled_location = "~/.python_search/t5_embeddings.pkl"
 
     def __init__(self):
         self._configuration = ConfigurationLoader().load_config()
+        self.model, self.tokenizer = T5Model().load_trained_model()
+
+
+    def get_embeddings(self, sentences: List[str]):
+        import torch
+
+        # Tokenize all sentences and convert to tensor format
+        inputs = self.tokenizer(sentences, padding=True, truncation=True, return_tensors='pt')
+
+        with torch.no_grad():
+            # Generate the outputs from the model
+            outputs = self.model.encoder(inputs['input_ids']).last_hidden_state
+
+        # Average the embeddings to get sentence-level embeddings
+        sentence_embeddings = torch.mean(outputs, dim=1)
+
+        return sentence_embeddings
+
 
     def save_missing_keys(self):
         """
@@ -27,16 +43,9 @@ class SemanticSearch:
         unique_bodies = []
 
         for key in unique_keys:
-            if key in entries:
-                body = str(entries[key])
-                print(f"For key '{key}', found body to encode: {body}")
-                unique_bodies.append(key + " " + body)
-            else:
-                print(f"Could not find body for key: {key}")
                 unique_bodies.append(key)
 
-        embeddings = to_embedding2(unique_bodies)
-        embeddings_np = [embedding.numpy() for embedding in embeddings]
+        embeddings = self.get_embeddings(unique_bodies)
 
         df_missing = pd.DataFrame()
         df_missing["key"] = unique_keys
@@ -66,35 +75,18 @@ class SemanticSearch:
                 unique_bodies.append(key)
 
 
-        embeddings = to_embedding2(unique_bodies)
-        embeddings_np = [embedding.numpy() for embedding in embeddings]
+        embeddings = [self.get_embeddings([body]) for body in unique_bodies]
 
         df = pd.DataFrame()
         df["key"] = unique_keys
         df['body'] = unique_bodies
-        df["embedding"] = embeddings_np
+        df["embedding"] = embeddings
 
         print("Pickled embeddings to: " + self.pickled_location)
         df.to_pickle(self.pickled_location)
 
     def _read_dataframe(self):
         return pd.read_pickle(self.pickled_location)
-
-    def rank_entries_by_query_similarity(self, query) -> List[str]:
-
-        from python_search.apps.notification_ui import send_notification
-        send_notification(f"Starting semantic search")
-        embedding_query = to_embedding([query])[0].numpy()
-
-        df = self._read_dataframe()
-        result = []
-        for i, key in enumerate(df['key'].tolist()):
-            similarity = cosine_similarity(embedding_query.reshape(1, -1), df.iloc[i]['embedding'].reshape(1, -1))
-            result.append((key, float(similarity)))
-
-        result = sorted(result, key=lambda x: x[1], reverse=True)
-
-        return [x[0] for x in result]
 
     def _get_missing_keys(self) -> List[str]:
         entries = self._configuration.commands
@@ -108,7 +100,7 @@ class SemanticSearch:
 
 def main():
     import fire
-    fire.Fire(SemanticSearch)
+    fire.Fire(T5Embeddings)
 
 if __name__ == "__main__":
     main()
